Application now compiles on Mac OSX (Yosemite).
Added `internalAppIsRunning` procedure for Darwin.

Uses `pgrep` instead of `ps` as MacOSX's ps lacks the `-C` option.
Instead we will use `pgrep` to query for the applications pid. If the
application is running we will get a positive integer (as a string) in
return which is returned as the `Result` after casting to `Int`.

Also: my first ever ObjectPascal code ;) Hooray!